refactored to add subfolders to config for routes and models
