+++ data/dbConfig.js
@@ -2,6 +2,6 @@ const knex = require('knex');
 require('dotenv').config();
 
 const dbEngine = process.env.DB || 'development';
-const config = require('../knexfile')[dbEngine]
+const config = require('../knexfile')
 
-module.exports = config;
+module.exports = knex(config[dbEngine]);
