+++ knexfile.js
@@ -3,13 +3,13 @@ require('dotenv').config()
 module.exports = {
   development: {
     client: 'sqlite3',
-    connection: { filename: './database/auth.db3' }, // change this if you want a different name for the database
+    connection: { filename: './data/auth.db3' }, // change this if you want a different name for the data
     useNullAsDefault: true, // used to avoid warning on console
     migrations: {
-      directory: './database/migrations',
+      directory: './data/migrations',
       tableName: 'dbmigrations',
     },
-    seeds: { directory: './database/seeds' },
+    seeds: { directory: './data/seeds' },
     pool: {
       afterCreate: (conn, done) => {
         conn.run('PRAGMA foreign_keys = ON', done)
